+++ packages/react-core/src/components/molecules/DatePicker/DatePicker.tsx
@@ -1,19 +1,25 @@
 import { format as formatDate } from 'date-fns';
 import * as React from 'react';
-import { useInputFocus } from '../../atoms/Input';
+import { ModalBaseProps, View } from 'react-native';
 import {
+  Hint,
+  InputContainerProps,
   PressableInputContainer,
-  PressableInputContainerProps,
-} from '../../atoms/PressableInputContainer';
+  useInputFocus,
+} from '../../atoms/Input';
 import { PressableSurfaceProps } from '../../atoms/PressableSurface';
-import { Text } from '../../atoms/Text';
-import { CalendarProps, DateRange, SelectionType } from '../Calendar';
+import { Text, TextProps } from '../../atoms/Text';
+import { Calendar, CalendarProps, DateRange, SelectionType } from '../Calendar';
 import { Modal } from './Modal';
+import { CalendarIcon, getStyledTextComponent } from './styled';
 
 export interface DatePickerProps<T extends SelectionType>
-  extends PressableInputContainerProps,
+  extends InputContainerProps,
     Omit<CalendarProps<T>, 'style'> {
   PressableElement?: React.FC<PressableSurfaceProps>;
+  TextComponent?: React.FC<TextProps>;
+  CalendarComponent?: React.FC<CalendarProps<T>>;
+  animationType?: ModalBaseProps['animationType'];
   placeholder?: string;
   onFocus?: () => void | never;
   onBlur?: () => void | never;
@@ -34,8 +40,20 @@ function DatePicker<T extends SelectionType>({
   onBlur,
   disabled,
   PressableElement,
+  hintComponent,
+  hint,
+  variant = 'default',
+  TextComponent = Text,
+  CalendarComponent = Calendar,
+  rightComponent,
+  animationType = 'fade',
+  style,
   ...rest
 }: DatePickerProps<T>): JSX.Element {
+  const _hint = hintComponent || (
+    <Hint TextComponent={TextComponent} text={hint} variant={variant} />
+  );
+
   const { focused, handleBlur, handleFocus } = useInputFocus(
     onFocus,
     onBlur,
@@ -66,25 +84,45 @@ function DatePicker<T extends SelectionType>({
     }
   };
 
+  const StyledText = getStyledTextComponent(TextComponent);
+
   return (
     <>
       {PressableElement ? (
         <PressableElement onPress={handlePressInput} />
       ) : (
-        <PressableInputContainer
-          onPress={handlePressInput}
-          focused={focused}
-          disabled={disabled}
-          {...rest}
-        >
-          <Text>{getDisplayValue()}</Text>
-        </PressableInputContainer>
+        <View style={style}>
+          <PressableInputContainer
+            onPress={handlePressInput}
+            focused={focused}
+            disabled={disabled}
+            LabelComponent={TextComponent}
+            variant={variant}
+            rightComponent={
+              <>
+                <CalendarIcon
+                  name="calendar-sharp"
+                  type="ionicon"
+                  size="centi"
+                />
+                {rightComponent}
+              </>
+            }
+            {...rest}
+          >
+            <StyledText fontWeight="bold" disabled={disabled}>
+              {getDisplayValue() || ' '}
+            </StyledText>
+          </PressableInputContainer>
+          {hint && _hint}
+        </View>
       )}
       <Modal
+        CalendarComponent={CalendarComponent}
         visible={modalVisible}
         onRequestClose={handleCloseModal}
         animated
-        animationType={'slide'}
+        animationType={animationType}
         month={month}
         year={year}
         onChange={onChange}