+++ packages/react-core/src/components/molecules/DatePicker/Modal.tsx
@@ -1,6 +1,6 @@
 import * as React from 'react';
 import { Modal as RNModal, ModalProps } from 'react-native';
-import { Calendar, CalendarProps, SelectionType } from '../Calendar';
+import { CalendarProps, SelectionType } from '../Calendar';
 import { Backdrop, ModalContent } from './styled';
 
 const Component = <T extends SelectionType>({
@@ -10,18 +10,21 @@ const Component = <T extends SelectionType>({
   month,
   year,
   onChange,
+  CalendarComponent,
   ...modalProps
-}: CalendarProps<T> & ModalProps) => {
+}: CalendarProps<T> &
+  ModalProps & { CalendarComponent: React.FC<CalendarProps<T>> }) => {
   return (
     <RNModal
-      {...modalProps}
-      animationType={'fade'}
       transparent
+      hardwareAccelerated
+      statusBarTranslucent
+      {...modalProps}
       onRequestClose={onRequestClose}
     >
-      <Backdrop onPress={onRequestClose}>
+      <Backdrop onPress={onRequestClose} effect="none">
         <ModalContent>
-          <Calendar
+          <CalendarComponent
             pointerEvents={'box-none'}
             type={type}
             value={value}