+++ packages/react-core/src/components/molecules/DatePicker/styled.ts
@@ -1,10 +1,18 @@
 import styled, { css } from '@emotion/native';
+import { FC } from 'react';
 import { Platform } from 'react-native';
 import { StyleProps } from '../../../types/defaults';
+import { Icon } from '../../atoms/Icon';
+import { disabledInputStyles, InputContainerProps } from '../../atoms/Input';
+import { TextProps } from '../../atoms/Text';
+import {
+  PressableSurface,
+  PressableSurfaceProps,
+} from '../../atoms/PressableSurface';
 
 const isWeb = Platform.OS === 'web';
 
-export const Backdrop = styled.Pressable<Partial<StyleProps>>`
+export const Backdrop = styled(PressableSurface)`
   ${isWeb
     ? css`
         justify-content: center;
@@ -32,4 +40,18 @@ export const ModalContent = styled.View<Partial<StyleProps>>`
     `}
   `}
   overflow: hidden;
+`;
+
+export const getStyledTextComponent = (component: FC<TextProps>) => {
+  return styled(component)(
+    (props: Partial<InputContainerProps> & Partial<StyleProps>) => css`
+      line-height: ${props.theme?.typography.h5.lineHeight};
+      ${disabledInputStyles(props)}
+    `
+  );
+};
+
+export const CalendarIcon = styled(Icon)<Partial<StyleProps>>`
+  padding: ${({ theme }) => theme.spacing.centi};
+  color: ${({ theme }) => theme.color.secondary.medium};
 `;